ext_packed_depth_stencil/readpixels-24_8: cast to int before abs()
The 1-bit depth tolerance check relies on abs() to allow value to be
either above or below expected by 1. With uint32_t operands, unsigned
subtraction wraps around when value < expected, producing a huge
positive result that abs() cannot recover from. Cast both operands
to int so the subtraction can be negative and abs() works as intended.

../tests/spec/ext_packed_depth_stencil/readpixels-24_8.c:63:6: warning: taking the absolute value of unsigned type 'uint32_t' (aka 'unsigned int') has no effect [-Wabsolute-value]

Fixes: db3e476c2f ("readpixels-24_8: use 1-bit tolerance for depth")
Suggested-by: Marek Olšák <maraeo@gmail.com>
Part-of: <https://gitlab.freedesktop.org/mesa/piglit/-/merge_requests/1132>

+++ tests/spec/ext_packed_depth_stencil/readpixels-24_8.c
@@ -60,7 +60,7 @@ test_pixel(int x, int y, uint32_t value)
 
 	/* Compare depth and stencil separately. Use 1-bit tolerance for depth. */
 	if ((value & 0xff) - (expected & 0xff) != 0 ||
-	    abs((value >> 8) - (expected >> 8)) > 1) {
+	    abs((int)(value >> 8) - (int)(expected >> 8)) > 1) {
 		fprintf(stderr,
 			"Expected 0x%08x at (%d,%d), found 0x%08x\n",
 			expected, x, y, value);
